refactor: 패키지 구조 변경 - infrastructure.s3 신설
