--- src/main/java/com/aicodinator/backend/domain/community/domain/constant/UploadType.java
@@ -1,6 +1,0 @@
-package com.aicodinator.backend.domain.community.domain.constant;
-
-public enum UploadType {
-    REPORT,
-    COMMUNITY_POST
-}
