+++ src/main/java/com/aicodinator/backend/domain/infrastructure/s3/domain/constant/UploadType.java
@@ -1,0 +1,6 @@
+package com.aicodinator.backend.domain.infrastructure.s3.domain.constant;
+
+public enum UploadType {
+    REPORT,
+    COMMUNITY_POST
+}
